Unit testing updates to fix mistakes
